requirements.txt
numpy==1.26.3
matplotlib==3.8.3
torch==2.3.1+cu118
pandas==2.2.3
dill==0.3.9
gym==0.26.2
eventlet==0.39.1
Flask==2.1.3
opencv-python==4.9.0.80
gymnasium==0.28.1
pygame==2.5.2
tqdm==4.66.4
pettingzoo==1.23.1
scipy==1.14.0
ipython==7.34.0
ipywidgets==8.1.2
imageio==2.34.2
numba==0.61.2
PyYAML==6.0.2
torchvision==0.18.1+cu118
torchaudio==2.3.1+cu118
setuptools==65.5.0
pillow==10.2.0